Refactoring code, added comments and code style changes for SpawnSystem
